requirements.txt
flask==3.0.0
gunicorn==21.2.0
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
pymysql==1.1.0
paramiko==3.4.0
pydub==0.25.1
edge-tts==7.2.7
google-generativeai==0.8.3
cryptography==44.0.0
